refactor: split LinkSection into StreakSection and BookmarkSection
